+++ packages/cmrjs/tests/cmr-utils/test-cmr-utils.js
@@ -33,6 +33,9 @@ const {
   granuleToCmrFileObject,
   mapFileEtags,
   removeEtagsFromFileObjects,
+  setECHO10Collection,
+  setUMMGCollection,
+  getCMRCollectionId,
 } = require('../../cmr-utils');
 const cmrUtil = rewire('../../cmr-utils');
 const { isCMRFile, isISOFile, getGranuleTemporalInfo } = cmrUtil;
@@ -643,7 +646,6 @@ test.serial('updateUMMGMetadata adds Type correctly to RelatedURLs for granule w
       bucketTypes,
       distributionBucketMap,
     });
-
     t.is(etag, expectedEtag, "ETag doesn't match");
     t.deepEqual(metadataObject.RelatedUrls.sort(sortByURL), expectedRelatedURLs.sort(sortByURL));
   } finally {
@@ -1335,11 +1337,15 @@ test(
       { short_name: 'sn2', version: '2' },
       { short_name: 'sn3', version: '3' },
     ];
-    const expected = { condition: { or: [
-      { and: [{ short_name: 'sn1' }, { version: '1' }] },
-      { and: [{ short_name: 'sn2' }, { version: '2' }] },
-      { and: [{ short_name: 'sn3' }, { version: '3' }] },
-    ] } };
+    const expected = {
+      condition: {
+        or: [
+          { and: [{ short_name: 'sn1' }, { version: '1' }] },
+          { and: [{ short_name: 'sn2' }, { version: '2' }] },
+          { and: [{ short_name: 'sn3' }, { version: '3' }] },
+        ],
+      },
+    };
 
     const actual = buildCMRQuery(results);
     t.deepEqual(actual, expected);
@@ -1351,4 +1357,97 @@ test('buildCMRQuery works with if the input results list is empty', (t) => {
   const expected = { condition: { or: [] } };
   const actual = buildCMRQuery(results);
   t.deepEqual(actual, expected);
+});
+
+test('setECHO10Collection updates echo10 collection name and version', async (t) => {
+  const filename = 'tests/cmr-utils/data/meta.xml';
+  const cmrObject = await promisify(xml2js.parseString)(fs.readFileSync(filename, 'utf-8'), xmlParseOptions);
+  const updated = setECHO10Collection(
+    cmrObject,
+    { name: 'a', version: 'b' }
+  );
+  t.is(updated.Granule.Collection.ShortName, 'a');
+  t.is(updated.Granule.Collection.VersionId, 'b');
+});
+
+test('setECHO10Collection fails when Echo10Files has no Granule field', (t) => {
+  const cmrObject = {};
+  t.throws(() => setECHO10Collection(cmrObject), {
+    message: 'invalid ECHO10 cmr metadata {}, must have granule tag',
+  });
+});
+
+test('setECHO10Collection adds collection if missing when Echo10Files has Granule object field', (t) => {
+  const cmrObject = { Granule: {} };
+  const updated = setECHO10Collection(
+    cmrObject, { name: 'a', version: 'b' }
+  );
+  t.is(updated.Granule.Collection.ShortName, 'a');
+  t.is(updated.Granule.Collection.VersionId, 'b');
+});
+
+test('setUMMGCollection updates umm meta object', (t) => {
+  const filename = 'tests/cmr-utils/data/ummg-meta.json';
+  const cmrObject = JSON.parse(fs.readFileSync(filename, 'utf-8'));
+  const updated = setUMMGCollection(cmrObject, { name: 'a', version: 'b' }, filename);
+
+  t.is(updated.CollectionReference.ShortName, 'a');
+  t.is(updated.CollectionReference.Version, 'b');
+});
+
+test('setUMMGCollection updates umm collection when missing', (t) => {
+  const cmrObject = {};
+  const updated = setUMMGCollection(cmrObject, { name: 'a', version: 'b' });
+  t.is(updated.CollectionReference.ShortName, 'a');
+  t.is(updated.CollectionReference.Version, 'b');
+});
+
+test('getCMRCollectionId gets umm collection', (t) => {
+  const cmrObject = {
+    CollectionReference: {
+      ShortName: 'a',
+      Version: '001',
+    },
+  };
+  t.is(getCMRCollectionId(cmrObject, 'a.cmr.json'), 'a___001');
+});
+
+test('getCMRCollectionId throws on invalid umm metadata', (t) => {
+  const cmrObject = {
+    CollectionReference: {
+      Version: '001',
+    },
+  };
+  t.throws(
+    () => getCMRCollectionId(cmrObject, 'a.cmr.json'),
+    {
+      message: 'UMMG metadata file has invalid collection configuration {"Version":"001"}',
+    }
+  );
+});
+
+test('getCMRCollectionId gets xml collection', (t) => {
+  const cmrObject = {
+    Granule: {
+      Collection: {
+        ShortName: 'a',
+        VersionId: '001',
+      },
+    },
+  };
+  t.is(getCMRCollectionId(cmrObject, 'a.cmr.xml'), 'a___001');
+});
+
+test('getCMRCollectionId throws on invalid xml metadata', (t) => {
+  const cmrObject = {
+    CollectionReference: {
+      Version: '001',
+    },
+  };
+  t.throws(
+    () => getCMRCollectionId(cmrObject, 'a.cmr.xml'),
+    {
+      message: 'ECHO10 metadata file has invalid collection configuration undefined',
+    }
+  );
 });