+++ CHANGELOG.md
@@ -6,6 +6,31 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- **CUMULUS-3751**
+  - Added `change-granule-collection-s3` to move granules to a different collection.
+    - expects a list of granuleIds along with a new (target) collection
+    - moves those granule files in S3 according to pathing of target collection
+    - update CMR metadata file according to new collection information
+  - Added CopyObject function in @cumulus/aws-client/S3 to facilitate multi-part s3 object copying
+  - Added functions to allow manipulation in memory of cmr metadata objects in @cumulus/cmrjs/cmr-utils
+    - updateUMMGMetadataObject updates file links for ummg metadata structure object
+    - updateEcho10XMLMetadataObject updates file links for echo10 metadata structure object
+    - setUMMGCollection sets collection name and version in ummg metadata structure object
+    - setEcho10Collection sets collection name and version in echo10 metadata structure object
+    - getCMRCollectionId gets collectionId from cmr metadata object using its filename to
+      determine how to correctly parse the object (echo10 vs ummg)
+  - Added MoveGranuleCollections workflow to cumulus core deployable according to terraform variables
+  - Added ingest module terraform variable "deploy_cumulus_workflows": a map of workflows that should be deployed
+    - as of merging only controls move_granule_collections_workflow
+    - defaults to true (deploy the workflow)
+
+### Fixed
+
+- **CUMULUS-3752**
+  - Fixed api return codes expected in api-client for bulkPatch and bulkPatchGranuleCollections
+
 ## [v18.5.3] 2025-01-21
 
 **Please note** changes in v18.5.3 may not yet be released in future versions, as this